work around PR#15127 and other tweaks
git-svn-id: https://svn.r-project.org/R/trunk@62192 00db46b3-68df-0310-9c12-caf00c1e9a41

+++ src/main/memory.c
@@ -342,7 +342,7 @@ static void init_gc_grow_settings()
 
     arg = getenv("R_GC_MEM_GROW");
     if (arg != NULL) {
-	int which = atof(arg);
+	int which = (int) atof(arg);
 	switch (which) {
 	case 0: /* very conservative -- the SMALL_MEMORY settings */
 	    R_NGrowIncrFrac = 0.0;
